zkvm: OpenVM 2 backend
The third zkVM backend, targeting OpenVM v2 (riscv32im-risc0-zkvm-elf,
transpiled to OpenVM's native format). It executes real Ethereum mainnet
blocks to the correct post-state root, passes the full precompile
golden-vector suite, and proves.

The design mirrors the ZisK backend — a Rust guest crate links
libmonad-zkvm-guest-openvm.a and supplies the eth-act runtime ABI — with one
difference that shapes most of the work: OpenVM has no eth-act runtime
library, so this crate provides everything. I/O (src/runtime.rs) and all 19
accelerators (src/accel.rs, src/bls12_map.rs), against OpenVM's guest
extensions where they exist and software otherwise. ripemd160, blake2f and
arbitrary-modulus modexp are pure software; the BLS12-381 maps and KZG point
evaluation are built on the accelerated field and pairing operations. The
shims follow the ZisK conventions throughout, so precompiles_impl.hpp needs
no per-backend branch and 1961/1961 golden vectors pass.

OpenVM is rv32im-only and the shared C++ tree was LP64 throughout, so most of
the non-zkvm diff is making it compile for ILP32. `unsigned __int128` does not
exist on 32-bit targets and accounted for ~85% of the initial errors; a
64x64->128 multiply and a 128/64 divide are the only operations needing a
wider type, so they are isolated as mul64x64 / div128by64 in
category/core/runtime/uint128.hpp behind MONAD_HAS_INT128, with
schoolbook-multiply and binary-long-division fallbacks. uint128_t's operators,
uint256/portable.hpp, knuth_div and storage_page.hpp's bitmap_t are all
written once against those two. Define MONAD_NO_INT128 to compile the
fallbacks on a 64-bit host.

46 sizeof/alignof/offsetof budgets across 14 headers hold only under the
64-bit host ABI, since every offset in a type containing a pointer or a size_t
moves on a 32-bit target; they now go through MONAD_ASSERT_LP64_LAYOUT
(category/core/config.hpp), which checks exactly as before on a 64-bit target
and compiles out on a 32-bit one. Restating each number for a second ABI would
have created a second set of magic constants that nothing keeps honest.
uint64_t-to-size_t narrowing is split by intent: hash values and provably
small indices take a plain cast, since truncating is the point, while lengths
deriving from block or transaction input go through a new narrow_to_size
(category/core/int.hpp), which asserts the value fits. And this toolchain's
newlib makes uint32_t `unsigned long` while size_t is `unsigned int` — same
width, distinct types — so untyped literals like `128ul` break std::min/max
deduction in ways x86 never shows; typed at each site.

Two build-side consequences, documented in place in zkvm/guest/CMakeLists.txt.
The toolchain needs an rv32im/ilp32 multilib, because -print-file-name
resolves per multilib and a wrong answer yields a silently mixed-ABI archive;
a configure-time ELF-class check now fails there rather than at the final
link, and build-riscv-toolchain.sh grows the multilib generator. libgcc
members are located by the symbol they define rather than by an ABI-dependent
file name, with a transitive closure over what those members themselves need.
The guest also builds with -fno-jump-tables on OpenVM only: GCC emits a switch
jump table as one rodata word per case holding case_target - table_base, an
R_RISCV_ADD32/R_RISCV_SUB32 pair on RISC-V, and powdr's ELF reader accepts
only R_RISCV_32 in data sections. It costs 0.019% fewer instructions here.

Every OpenVM crate resolves to github.com/powdr-labs/openvm.git tag
v2.0.0-beta.2-powdr.1 rather than openvm-org upstream. That is the tag
powdr-openvm pins, and this backend is meant to feed powdr's autoprecompiles.
The version has to be the same in the guest, the host driver and the
cargo-openvm that transpiles, because the .vmexe serialisation is not
compatible with upstream v2.0.1. SP1 and this OpenVM cannot share a lockfile
(SP1 pins p3-field =0.4.3-succinct against =0.4.1 here), so build-support's
sp1-build dependency moves behind an `sp1` feature that only zkvm/sp1/script
enables.

zkvm/openvm/script is the host driver, for the things cargo-openvm cannot
express: framing the witness as one write_bytes vector, raising
num_public_values, and proving. It executes, meters and proves, splitting core
proving from recursion so the two can be compared separately. A second binary
behind a `powdr` feature runs the same guest through powdr's autoprecompile
pipeline, sharing the phase-split prove so both sets of numbers come from one
definition. BLOCKHASH resolves from the witness's ancestor headers, which
arrive in exactly the order BlockHashBufferFinalized::set requires.

Mainnet blocks 22200001-22200005 all reproduce their expected post-state root,
in 1.9-7.2 s of interpreted execution each. Block 22200003 proves and verifies
end to end in 1 h 48 m at 48.5 GiB peak, or 1 h 12 m with 16 autoprecompiles —
1.51x, and the corpus floor rather than a typical figure, since segment counts
across the other four blocks improve by 1.59-1.70x. Core proving is ~90% of
the cost and OpenVM proves segments strictly sequentially, so per-block
latency is linear in segment count. Numbers and their caveats are in
zkvm/openvm/README.md.

+++ category/core/address.hpp
@@ -118,8 +118,9 @@ struct std::hash<monad::Address>
 {
     size_t operator()(monad::Address const &x) const noexcept
     {
-        return ankerl::unordered_dense::detail::wyhash::hash(
-            x.bytes, sizeof(x.bytes));
+        return static_cast<size_t>(
+            ankerl::unordered_dense::detail::wyhash::hash(
+                x.bytes, sizeof(x.bytes)));
     }
 };
